+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Form->validate and return errors to base controllers
 - Added Leaf Veins path replace to base controller
 - Added Leaf\Core\Str: equivalent of Illuminate\Support\Str with added methods
+- Finished Mysqli and started PDO
 
 
 #### Fixed